refactor(internal): Move fsStreamReady to internal package
