+++ packages/internal/src/utils.js
@@ -228,7 +228,7 @@ export const makeMeasureSeconds = currentTimeMillisec => {
  * @param {Error[]} errors
  * @param {string} [message]
  */
-const makeAggregateError = (errors, message) => {
+export const makeAggregateError = (errors, message) => {
   const err = new Error(message);
   Object.defineProperties(err, {
     name: {
@@ -279,3 +279,39 @@ export const PromiseAllOrErrors = async values => {
         )
         .then(error => Promise.reject(error)),
   );
+
+/**
+ * @param {import("fs").ReadStream | import("fs").WriteStream} stream
+ * @returns {Promise<void>}
+ */
+export const fsStreamReady = stream =>
+  new Promise((resolve, reject) => {
+    if (stream.destroyed) {
+      reject(new Error('Stream already destroyed'));
+      return;
+    }
+
+    if (!stream.pending) {
+      resolve();
+      return;
+    }
+
+    const onReady = () => {
+      cleanup(); // eslint-disable-line no-use-before-define
+      resolve();
+    };
+
+    /** @param {Error} err */
+    const onError = err => {
+      cleanup(); // eslint-disable-line no-use-before-define
+      reject(err);
+    };
+
+    const cleanup = () => {
+      stream.off('ready', onReady);
+      stream.off('error', onError);
+    };
+
+    stream.on('ready', onReady);
+    stream.on('error', onError);
+  });
